requirements.txt
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
earthengine-api==1.7.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
numpy==2.3.5
pandas==2.3.3
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
